Conflicts:
	.gitignore
	spec/models/contest_spec.rb
	spec/models/match_spec.rb
	spec/models/player_match_spec.rb
	spec/models/player_spec.rb
	spec/models/referee_spec.rb

+++ db/migrate/20131030135037_create_player_matches.rb
@@ -1,7 +1,7 @@
 class CreatePlayerMatches < ActiveRecord::Migration
   def change
     create_table :player_matches do |t|
-      t.double :score
+      t.float :score
       t.string :result
       t.references :player, index: true
       t.references :match, index: true
